Create a JobFactory
