+++ src/BigQueryBundle/BigQuery/UnitOfWork.php
@@ -16,32 +16,25 @@ class UnitOfWork
     private $bigQueryClient;
 
     /**
-     * @var FileSystemInterface
-     */
-    private $fileSystem;
-
-    /**
-     * @var string
-     */
-    private $bucket;
-
-    /**
      * @var MetadataRepository
      */
     private $metadataRepository;
+
+    /**
+     * @var JobFactory
+     */
+    private $jobFactory;
 
     private $data = [];
 
     public function __construct(
         \Google_Service_Bigquery $bigQueryClient,
-        FileSystemInterface $fileSystem,
-        $bucket,
-        MetadataRepository $metadataRepository
+        MetadataRepository $metadataRepository,
+        JobFactory $jobFactory
     ) {
         $this->bigQueryClient = $bigQueryClient;
-        $this->fileSystem = $fileSystem;
-        $this->bucket = $bucket;
         $this->metadataRepository = $metadataRepository;
+        $this->jobFactory = $jobFactory;
     }
 
     public function addData(RowInterface $data)
@@ -76,64 +69,7 @@ class UnitOfWork
 
     private function uploadData(MetadataInterface $metadata, array $data)
     {
-        $job = new \Google_Service_Bigquery_Job();
-
-        $name = 'reporting-' . date('Y-m-d') . '-' . uniqid() . '.json';
-
-        /*
-         * When importing data into bigquery, via a json file, null fields must be omitted. Otherwise it'll try to treat
-         * this field as the string "null".
-         * @see {https://stackoverflow.com/a/32619240}
-         * So we need to explicitly call jsonSerialize and strip null values in this array
-         */
-        array_walk(
-            $data,
-            function(\JsonSerializable &$item) {
-                $item = array_filter(
-                    $item->jsonSerialize(),
-                    function ($value) {
-                        return $value !== null;
-                    }
-                );
-            }
-        );
-
-        // We create a file on google cloud storage and then leave it here, so we can inspect it if needed.
-        $this->fileSystem->store(
-            $this->bucket,
-            $name,
-            'application/json',
-            implode(PHP_EOL, array_map('json_encode', $data))  // NewLine delimited JSON
-        );
-
-        $jobConfiguration = new \Google_Service_Bigquery_JobConfiguration();
-        $jobConfigurationLoad = new \Google_Service_Bigquery_JobConfigurationLoad();
-        $jobConfigurationLoad->setSourceUris(['gs://' . $this->bucket . '/' . $name]);
-
-        $reference = new \Google_Service_Bigquery_TableReference();
-        $reference->setDatasetId($metadata->getDatasetId());
-        $reference->setTableId($metadata->getTableId());
-        $reference->setProjectId($metadata->getProjectId());
-
-        $jobConfigurationLoad->setDestinationTable($reference);
-        $jobConfigurationLoad->setSourceFormat("NEWLINE_DELIMITED_JSON");
-        $jobConfigurationLoad->setWriteDisposition('WRITE_APPEND');
-
-        $schema = new \Google_Service_Bigquery_TableSchema();
-        $schema->setFields($metadata->getSchema());
-        $jobConfigurationLoad->setSchema($schema);
-
-        /**
-         * We create a partition per day, using the column "date" from our dataset.
-         */
-        $timePartitioning = new \Google_Service_Bigquery_TimePartitioning();
-        $timePartitioning->setType('DAY');
-        $timePartitioning->setField('date');
-        $jobConfigurationLoad->setTimePartitioning($timePartitioning);
-        $jobConfiguration->setLoad($jobConfigurationLoad);
-
-        $job->setConfiguration($jobConfiguration);
-
+        $job = $this->jobFactory->createJob($metadata, $data);
         $this->bigQueryClient->jobs->insert($metadata->getProjectId(), $job);
     }
 }
